Create optional in-place word replacement
For the following use case: Inserting a custom <span></span> tag (or any html tag) around a keyword. I want the replacement to use the original sentence word, not the keyword. If I am doing case insensitive matching (most of the time), I want the inserted tag to preserve the original word. 

With the current implementation, this is impossible to accomplish without switching to case sensitive matching. This small intervention allows for custom replacement without affecting anything else.